Add color widget
