+++ enterprise/frontend/src/metabase-enterprise/whitelabel/components/ColorSchemeWidget/ColorSchemeWidget.styled.tsx
@@ -3,5 +3,6 @@ import { color } from "metabase/lib/colors";
 
 export const SectionTitle = styled.div`
   color: ${color("text-medium")};
+  font-weight: bold;
   margin-bottom: 1rem;
 `;
